+++ admin/language/english/total/mollie_payment_fee.php
@@ -9,9 +9,25 @@ $_['text_success']            = "Success: You have successfully modified your Mo
 $_['text_enable']             = "Enable";
 $_['text_disable']            = "Disable";
 
+// Tab
+$_['tab_general']            = "General";
+$_['tab_charge']             = "Charges";
+
 // Entry
 $_['entry_status']	 		= "Status";
 $_['entry_sort_order']      = "Sort Order";
+$_['entry_tax_class']       = "Tax Class";
+$_['entry_title']           = "Title";
+$_['entry_payment_method']  = "Payment Method";
+$_['entry_cost']            = "Cost / Fee (Ex: 5, 5%)";
+$_['entry_store']           = "Store";
+$_['entry_customer_group']  = "Customer Group";
+$_['entry_geo_zone']        = "Geo Zone";
+$_['entry_priority']        = "Priority";
+
+// Button
+$_['button_add_charge']        = "Add Charge";
+$_['button_remove_charge']     = "Remove Charge";
 
 // Error
 $_['error_permission']      = "Warning: You don't have permission to modify the Mollie module.";