+++ admin/language/french/total/mollie_payment_fee.php
@@ -9,9 +9,25 @@ $_['text_success']            = "Succès: Vous avez réussi à modifier les para
 $_['text_enable']             = "Activer";
 $_['text_disable']            = "Désactiver";
 
+// Tab
+$_['tab_general']            = "Général";
+$_['tab_charge']             = "Des charges";
+
 // Entry
 $_['entry_status']	 		= "Status";
 $_['entry_sort_order']      = "Ordre de triage";
+$_['entry_tax_class']       = "Tax Class";
+$_['entry_title']           = "Titre";
+$_['entry_payment_method']  = "Mode de paiement";
+$_['entry_cost']            = "Coût / Frais (Ex: 5, 5%)";
+$_['entry_store']           = "Boutique";
+$_['entry_customer_group']  = "Groupe de clients";
+$_['entry_geo_zone']        = "Zone géographique";
+$_['entry_priority']        = "Priorité";
+
+// Button
+$_['button_add_charge']        = "Add Charge";
+$_['button_remove_charge']     = "Remove Charge";
 
 // Error
 $_['error_permission']      = "Attention: Vous n'avez pas l'autorisation de modifier les méthodes de paiement Mollie.";