+++ admin/language/german/total/mollie_payment_fee.php
@@ -9,9 +9,25 @@ $_['text_success']            = "Erfolg: Die Einstellungen für dieses Modul wur
 $_['text_enable']             = "Aktivieren";
 $_['text_disable']            = "Deaktivieren";
 
+// Tab
+$_['tab_general']            = "Allgemein";
+$_['tab_charge']             = "Gebühren";
+
 // Entry
 $_['entry_status']	 		= "Status";
 $_['entry_sort_order']      = "Sortierreihenfolge";
+$_['entry_tax_class']       = "Tax Class";
+$_['entry_title']           = "Titel";
+$_['entry_payment_method']  = "Zahlungsmethode";
+$_['entry_cost']            = "Kosten / Gebühr (Ex: 5, 5%)";
+$_['entry_store']           = "Geschäft";
+$_['entry_customer_group']  = "Kundengruppe";
+$_['entry_geo_zone']        = "Geo Zone";
+$_['entry_priority']        = "Priorität";
+
+// Button
+$_['button_add_charge']        = "Add Charge";
+$_['button_remove_charge']     = "Remove Charge";
 
 // Error
 $_['error_permission']      = "Warnung: Sie haben keine Berechtigung, das Modul zu bearbeiten.";