+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 # Changelog #
 
+#### Changes in release 11.1.0
+  + Updated mollie payment details interface on order page
+  + Removed mollie payment fee options from 'payment' module and added them to 'order total' module (Mollie Payment Fee)
+  + Updated mollie payment fee interface and added more options
+  + Updated mollie api version
+  + Updated module design
+  + Removed mollie connect feature
+  + Bugfix and improvements
+
 #### Changes in release 11.0.1
   + Fixed payment failed status
   + Fixed permission error in admin after fresh install of the module